Remove export_mix_env function
The `export_mix_env` function relies upon a `MIX_ENV` being set in the
`$env_path`. If no `MIX_ENV` is set, the buildpack defaults to prod.

Because we are always setting a `MIX_ENV` in our application manifest,
we don't need to rely upon this fallback. This commit removes the
`export_mix_env` function and requires users to specify a `MIX_ENV`.